refactor: bifurcates Utils into specific domains
