+++ test/lastfm_archive/livebook_test.exs
@@ -7,7 +7,8 @@ defmodule LastfmArchive.LivebookTest do
   alias LastfmArchive.Cache
   alias LastfmArchive.Cache.Server, as: CacheServer
   alias LastfmArchive.Livebook, as: LFM_LB
-  alias LastfmArchive.Utils
+
+  import LastfmArchive.Utils.Archive, only: [user_dir: 1]
 
   @cache :livebooktest_cache
 
@@ -18,7 +19,7 @@ defmodule LastfmArchive.LivebookTest do
     {:ok, _pid} = start_supervised(%{id: @cache, start: {CacheServer, :start_link, [[name: @cache]]}})
     cache = %{{"a_user", 2006} => %{{1_138_752_000, 1_138_838_399} => {33, [:ok]}}}
 
-    cache_file = "#{Utils.user_dir("a_user")}/#{Cache.cache_dir()}/2006"
+    cache_file = "#{user_dir("a_user")}/#{Cache.cache_dir()}/2006"
     stub_with(LastfmClient.impl(), LastfmArchive.LastfmClientStub)
     LastfmArchive.PathIOMock |> stub(:wildcard, fn _, _ -> [cache_file] end)
     LastfmArchive.FileIOMock |> stub(:read!, fn _ -> cache[{"a_user", 2006}] |> :erlang.term_to_binary() end)
